requirements.txt
torchtext
nltk
spacy
gpytorch
networkx
pandas
numpy
scipy
pandas
dask
ipython
tqdm
click
matplotlib
seaborn
altair
statsmodels
scikit-learn
dill
spacy
seaborn
simplejson
transformers
sklearn
sdeint

# for development only (could also be kept in a separate environment file)
jupyterlab
pytest
pytest-cov
tox
pre_commit
nbdime
nbstripout
sphinx
recommonmark
GitPython
